+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
-- * 193d8cd - 2018-04-21: Release v9.3.6 
+- * 1f904d1 - 2018-04-25: Fixed issue #202 - Torrent client delaying load times on all pages 
+- * ee7b8bc - 2018-04-21: Release v9.3.6 
 - * c32702b - 2018-04-18: Daily/Backlog searches now skip downloaded episodes with higher quality then show&#x27;s set preferred or best. Daily searches skip downloaded episodes if archive on first match enabled. 
 - * 9e76bc0 - 2018-04-15: Manual searches and Removal of shows now take extreme priority in queue 
 - * e8cea87 - 2018-04-15: Manual searches and Removal of shows now take extreme priority in queue